[UPDATE] do not return an error from NewSMILeaf - just log a warning.
Return error from NewSMILeaf made the boilerplate of building up a tree
too much to bear, and meant that the SMI tree functions couldn't be
composed.

Although passing an invalid type over the pass_persist line protocol is
an error, it's not really an error that exists at the time that a leaf
is instantiated in memory.